dev-requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.9
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
copier==9.8.0
coverage==7.9.2
Deprecated==1.2.14
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
dunamai==1.25.0
email_validator==2.2.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.2
filelock==3.18.0
funcy==2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-ansible-filters==1.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
opentelemetry-api==1.34.1
opentelemetry-distro==0.55b1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
packaging==25.0
pathlib2==2.3.7.post1
pathspec==0.12.1
pipdeptree==2.26.1
platformdirs==4.3.8
pluggy==1.6.0
plumbum==1.9.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
protobuf==5.29.5
py==1.11.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyright==1.1.403
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
questionary==2.1.0
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.5.1
ruff==0.12.2
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stomp.py==8.2.0
tox==3.28.0
tox-direct==0.4
typer==0.16.0
types-mock==5.2.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
